Feat: create BookRepo

+++ src/main/java/io/spring/training/boot/server/services/BookService.java
@@ -1,6 +1,7 @@
 package io.spring.training.boot.server.services;
 
 import io.spring.training.boot.server.models.Book;
+import io.spring.training.boot.server.repositories.BookRepo;
 import lombok.AllArgsConstructor;
 import lombok.Lombok;
 import org.springframework.stereotype.Service;
@@ -16,9 +17,8 @@ public class BookService {
         return bookRepo.save(book);
     }
 
-    public Book findBookById(int id){
+    public Book findBookById(long id){
         Optional<Book> book = bookRepo.findById(id);
-
         return book.orElse(new Book());
     }
 }
